+++ src/array/everyPass.js
@@ -1,6 +1,6 @@
 import _curry2 from '../_internals/_curry2'
-import reduce from '../function/reduce'
-import reduced from '../function/reduced'
+import _reduce from '../_internals/_reduce'
+import _reduced from '../_internals/_reduced'
 
 /**
  * @name everyPass
@@ -28,6 +28,6 @@ import reduced from '../function/reduced'
  * fn(2) // => true
  * fn(0) // => false
  */
-const everyPass = (fns, data) => reduce((f, acc) => !f(data) ? reduced(false) : acc, true, fns)
+const everyPass = (fns, data) => _reduce((f, acc) => !f(data) ? _reduced(false) : acc, true, fns)
 
 export default _curry2(everyPass)